requirements.txt
keybert==0.7.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
python-dotenv==0.21.1
streamlit==1.17.0
transformers==4.24.0
tweepy==4.12.1
